requirements.txt
sqlalchemy
psycopg2
cassandra-driver
Flask
strawberry-graphql[debug-server]
